:bug: fix unclean screen at start and end
Before, At-Man started on a screen that still displayed what was on the
console screen when untitled-rl started up. The prompt after quitting
overlaps the now even dirtier screen.

Now, the screen is automatically cleaned at start of game and cleaned
again with a color reset at end of game loop.